Skill tweak + Wisp fix
Tweaked range of most skills to be either +5 range for melee or -5 for ranged/magic.  Realistically, gives player a disadvantage to a fair fight at ranged but advantage to melee.  Fixes loading a spell though, and your character stopping at max distance instead of shooting.
Wisp fix shouldn't be noticeable.